project/requirements.txt
# loads necessary flask packages
Flask
Flask-Cors
Flask-WTF # forms
Flask-Migrate # required for migrating database with any changes
flask-login # required for password verification
Flask-SQLAlchemy # required for SQLite
email-validator
sqlalchemy.orm

# mysql-connector-python # required for MySQL connection
